# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API for managing contacts information",
    "version": "1.0.0"
  },
  "host": "cse-341-1lpa.onrender.com",
  "basePath": "/",
  "tags": [
    {
      "name": "contacts",
      "description": "Endpoints for managing contacts"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/debug/db": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Phillip"
        },
        "lastName": {
          "type": "string",
          "example": "Mubbala"
        },
        "email": {
          "type": "string",
          "example": "mubbalaphillip@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "yellow"
        },
        "birthday": {
          "type": "string",
          "example": "2003-02-28"
        }
      }
    },
    "ContactInput": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Rinah"
        },
        "lastName": {
          "type": "string",
          "example": "Mutesi"
        },
        "email": {
          "type": "string",
          "example": "mutesirinahl@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "Purple"
        },
        "birthday": {
          "type": "string",
          "example": "2005-11-14"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "favoriteColor",
        "birthday"
      ]
    },
    "ContactResponse": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "682274ddd0a5b29bade88762"
        },
        "firstName": {
          "type": "string",
          "example": "Andrew"
        },
        "lastName": {
          "type": "string",
          "example": "Samuku"
        },
        "email": {
          "type": "string",
          "example": "samukuandrew@gmail.com"
        },
        "favoriteColor": {
          "type": "string",
          "example": "blue"
        },
        "birthday": {
          "type": "string",
          "example": "2000-01-21"
        }
      }
    },
    "NewContactResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "682274ddd0a5b29bade88762"
        },
        "message": {
          "type": "string",
          "example": "Contact created successfully"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Error message"
        },
        "message": {
          "type": "string",
          "example": "Detailed error information"
        }
      }
    }
  }
}